fixed logic create jwt
