+++ .env.example
@@ -5,3 +5,4 @@ DB_PASSWORD=
 DB_PORT=
 DB_DIALECT
 SENDGRID_API_KEY=
+JWT_KEY=
